requirements.txt
telebot==0.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0